Split What's New e2e spec into focused files
ref https://github.com/TryGhost/Ghost/actions/workflows/ci.yml

Draft shard-balancing split so Playwright can schedule the banner, modal, and badge coverage independently.
